# Conflicts:
#	niryo_robot_bringup/launch/niryo_robot_base_common.launch.xml
#	niryo_robot_hardware_stack/end_effector_interface/src/end_effector_interface_core.cpp
#	niryo_robot_hardware_stack/joints_interface/include/joints_interface/joint_hardware_interface.hpp
#	niryo_robot_hardware_stack/joints_interface/src/joints_interface_core.cpp
#	niryo_robot_hardware_stack/ttl_driver/CMakeLists.txt
#	niryo_robot_hardware_stack/ttl_driver/src/abstract_dxl_driver.cpp
#	niryo_robot_hardware_stack/ttl_driver/src/abstract_end_effector_driver.cpp
#	niryo_robot_hardware_stack/ttl_driver/src/abstract_motor_driver.cpp
#	niryo_robot_hardware_stack/ttl_driver/src/abstract_stepper_driver.cpp
#	niryo_robot_hardware_stack/ttl_driver/src/ttl_interface_core.cpp

+++ niryo_robot_hardware_stack/ttl_driver/include/ttl_driver/xl430_reg.hpp
@@ -26,7 +26,7 @@ struct XL430Reg
 {
     static constexpr common::model::EHardwareType motor_type = common::model::EHardwareType::XL430;
 
-    static constexpr int PROTOCOL_VERSION                       = 2.0;
+    static constexpr float PROTOCOL_VERSION                     = 2.0;
     static constexpr int MODEL_NUMBER                           = 1060;
     static constexpr int VOLTAGE_CONVERSION                     = 10;
